move create tag list functionality to dom updates
