+++ src/recipe-repo.js
@@ -28,20 +28,5 @@ export default class RecipeRepo {
     return filteredRecipes;
   }
 }
-  // filterRecipesByTag(selected, element) {
-  //   const filteredRecipes = [];
-  //   selected.forEach(tag => {
-  //   const allRecipes = this.recipes
-  //   .filter(recipe => recipe.tags.includes(tag))
-  //   .forEach(recipe => {
-  //     if (!filteredRecipes.includes(recipe)) {
-  //       filteredRecipes.push(recipe)
-  //     };
-  //   });
-  //   }) 
-  //   return filteredRecipes
-  //   }
-  // } 
-
